Keep hidden items (#1377)
* When items are hidden add them to a DiffContext list

* no need to keep m_nHiddenItems since we now have a list of the currently hidden items.

* remove two introduced warnings

* .winmerge stores a list of hidden items

* hidde-item in singular (not plural)

* saveHiddenItems inside <path> node

* Fixing indentation to use Eric Allman style

* load the hiddenItems from the .winmerge project file + keep an use the token

* load hidden-items from .winmerge and hidding some test items

* https://github.com/WinMerge/winmerge/pull/1377#issuecomment-1164375545

* Add "Hidden Items" to the project options

* Update ChangeLog.md

* revert ChangeLog.md and README.md, applying requested changes

* rever ChangeLog.md empty space

* https://github.com/WinMerge/winmerge/pull/1377#discussion_r912328299

* update ChangeLog.md

+++ Src/DirView.cpp
@@ -90,7 +90,6 @@ IMPLEMENT_DYNCREATE(CDirView, CListView)
 
 CDirView::CDirView()
 		: m_pList(nullptr)
-		, m_nHiddenItems(0)
 		, m_compareStart(0)
 		, m_bTreeMode(false)
 		, m_dirfilter(std::bind(&COptionsMgr::GetBool, GetOptionsMgr(), _1))
@@ -2541,6 +2540,8 @@ LRESULT CDirView::OnUpdateUIMessage(WPARAM wParam, LPARAM lParam)
 			else
 				Redisplay();
 		}
+
+		HideItems(GetDiffContext().m_vCurrentlyHiddenItems);
 	}
 
 	return 0; // return value unused
@@ -3264,15 +3265,64 @@ void CDirView::OnUpdateItemRename(CCmdUI* pCmdUI)
  */
 void CDirView::OnHideFilenames()
 {
+	CDiffContext& ctxt = GetDiffContext();
+	int selection_index;
+	String hiddden_item_path;
+	
 	m_pList->SetRedraw(FALSE);	// Turn off updating (better performance)
 	DirItemIterator it;
+
 	while ((it = SelRevBegin()) != SelRevEnd())
 	{
 		DIFFITEM &di = *it;
+		selection_index = it.m_sel;
+		hiddden_item_path = di.getItemRelativePath();
 		SetItemViewFlag(di, ViewCustomFlags::HIDDEN, ViewCustomFlags::VISIBILITY);
-		DeleteItem(it.m_sel);
-		m_nHiddenItems++;
+		DeleteItem(selection_index);
+		ctxt.m_vCurrentlyHiddenItems.push_back(hiddden_item_path);
 	}
+	m_pList->SetRedraw(TRUE);	// Turn updating back on
+}
+
+/**
+ * @brief determine if an item-relative-path is contained in the list of items to hide
+ */
+bool CDirView::IsItemToHide(const String& currentItem, const std::vector<String>& ItemsToHide) const
+{
+	return std::find(ItemsToHide.begin(), ItemsToHide.end(), currentItem) != ItemsToHide.end();
+}
+
+/**
+ * @brief hides items specified in the .winmerge file
+ */
+
+void CDirView::HideItems(const std::vector<String>& ItemsToHide)
+{
+	CDiffContext& ctxt = GetDiffContext();
+	DIFFITEM *diffpos = ctxt.GetFirstDiffPosition();
+	while (diffpos != nullptr)
+	{
+		DIFFITEM &di = ctxt.GetNextDiffRefPosition(diffpos);
+		if (IsItemToHide(di.getItemRelativePath(), ItemsToHide))
+			SetItemViewFlag(di, ViewCustomFlags::HIDDEN, ViewCustomFlags::VISIBILITY);
+	}
+
+	m_pList->SetRedraw(FALSE);	// Turn off updating (better performance)
+
+	int num_hidden = 0;
+	const size_t num_to_hide = ItemsToHide.size();
+	DirItemIterator it = RevBegin();
+	while((num_hidden < num_to_hide) && (it != RevEnd()))
+	{
+		DIFFITEM& di = *it;
+		if (di.customFlags & ViewCustomFlags::HIDDEN)
+		{
+			DeleteItem(it.m_sel);
+			num_hidden++;
+		}
+		++it;
+	}
+
 	m_pList->SetRedraw(TRUE);	// Turn updating back on
 }
 
@@ -3495,8 +3545,9 @@ void CDirView::OnUpdateStatusNum(CCmdUI* pCmdUI)
  */
 void CDirView::OnViewShowHiddenItems()
 {
+	CDiffContext& ctxt = GetDiffContext();
 	SetItemViewFlag(GetDiffContext(), ViewCustomFlags::VISIBLE, ViewCustomFlags::VISIBILITY);
-	m_nHiddenItems = 0;
+	ctxt.m_vCurrentlyHiddenItems.clear();
 	Redisplay();
 }
 
@@ -3505,7 +3556,8 @@ void CDirView::OnViewShowHiddenItems()
  */
 void CDirView::OnUpdateViewShowHiddenItems(CCmdUI* pCmdUI)
 {
-	pCmdUI->Enable(m_nHiddenItems > 0);
+	const CDiffContext& ctxt = GetDiffContext();
+	pCmdUI->Enable(ctxt.m_vCurrentlyHiddenItems.size() > 0);
 }
 
 /**
@@ -4148,6 +4200,8 @@ void CDirView::OnSearch()
 	CDirDoc *pDoc = GetDocument();
 	m_pList->SetRedraw(FALSE);	// Turn off updating (better performance)
 	int nRows = m_pList->GetItemCount();
+	CDiffContext& ctxt = GetDiffContext();
+
 	for (int currRow = nRows - 1; currRow >= 0; currRow--)
 	{
 		DIFFITEM *pos = GetItemKey(currRow);
@@ -4157,6 +4211,7 @@ void CDirView::OnSearch()
 		bool bFound = false;
 		DIFFITEM &di = GetDiffItem(currRow);
 		PathContext paths;
+
 		for (int i = 0; i < pDoc->m_nDirs; i++)
 		{
 			if (di.diffcode.exists(i) && !di.diffcode.isDirectory())
@@ -4193,9 +4248,10 @@ void CDirView::OnSearch()
 		}
 		if (!bFound)
 		{
+			String hiddden_item_path = di.getItemRelativePath();
 			SetItemViewFlag(di, ViewCustomFlags::HIDDEN, ViewCustomFlags::VISIBILITY);
 			DeleteItem(currRow);
-			m_nHiddenItems++;
+			ctxt.m_vCurrentlyHiddenItems.push_back(hiddden_item_path);
 		}
 	}
 	m_pList->SetRedraw(TRUE);	// Turn updating back on
